Move `getDrops` to the mixed slab hooks
